게시물 페이지 박스

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -10,7 +10,7 @@
 <body>
 	<h1>게시물 상세화면</h1>
 		<a href="./list">목록</a>
-		<a href="./doDelete?num=${article.num}">삭제</a>
+		<a onclick="if( confirm('삭제하시겠습니까?') == false) return false;" href="./doDelete?num=${article.num}">삭제</a>
 		<a href="./modify?num=${article.num}"> 수정</a>
 		<div> 게시물 번호 : ${article.num}</div>
 		<div> 게시물 작성일 : ${article.regDate}</div>
